Code structure refactoring
