+++ requirements.txt
@@ -1,0 +1,4 @@
+requests>=2.31.0
+python-dotenv>=1.0.0
+langchain-groq>=0.1.0
+langchain>=0.1.0
